Tidy the Stub Source
To the module root because there aren't any other elements to export from
this module.
